# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Estados",
        "description": "API para gestión de estados",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://uptyab.test/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Obtener lista de roles",
                "operationId": "77e3636d8592997ce9d6d27b50857a69",
                "responses": {
                    "200": {
                        "description": "Lista de roles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Role"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Crear un nuevo rol",
                "operationId": "4a563842fbfcad237f63da43065fca3d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rol",
                                    "descripcion",
                                    "nivel_acceso"
                                ],
                                "properties": {
                                    "rol": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "Administrador"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "maxLength": 500,
                                        "example": "Tiene acceso completo al sistema"
                                    },
                                    "nivel_acceso": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rol creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/roles/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Obtener un rol por ID",
                "operationId": "7fb0667a295b7678a30353aa11739683",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rol encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Actualizar un rol existente",
                "operationId": "786d6990093a8eb3a56931f19a827e43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "rol": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "Administrador"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "maxLength": 500,
                                        "example": "Tiene acceso completo al sistema"
                                    },
                                    "nivel_acceso": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rol actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Eliminar un rol",
                "operationId": "8285c46981c60ad600aa4077307aec04",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Rol eliminado exitosamente"
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Role": {
                "title": "Role",
                "description": "Modelo que representa un rol en el sistema",
                "properties": {
                    "id": {
                        "description": "ID del rol",
                        "type": "integer",
                        "format": "int64"
                    },
                    "rol": {
                        "description": "Nombre del rol",
                        "type": "string",
                        "maxLength": 50
                    },
                    "descripcion": {
                        "description": "Descripción detallada del rol",
                        "type": "string",
                        "maxLength": 500
                    },
                    "nivel_acceso": {
                        "description": "Nivel de acceso del rol (mayor número = más privilegios)",
                        "type": "integer"
                    },
                    "estado": {
                        "description": "Estado del rol (activo/inactivo)",
                        "type": "boolean"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}